fix: 修改oeta基础报表获取广告转化数据

+++ co-driver-statistics/co-driver-statistics-services/co-driver-statistics-service-ads/src/main/java/com/leo/ad/codriver/ads/entity/AdsDailyOetaBaseReport.java
@@ -48,7 +48,7 @@ public class AdsDailyOetaBaseReport implements BaseEntity {
     private Long adDirectSoldShowNum;
     private BigDecimal avgUserAdDirectSoldShowNum;
     private Long adDirectSoldClickNum;
-    private Long adDirectSoldClickShowRate;
+    private BigDecimal adDirectSoldClickShowRate;
     private BigDecimal adDirectSoldRate;
     private BigDecimal avgUserAdShowNum;
     private BigDecimal avgUserAdInterShowNum;
